test: 테스트 코드 개선
- `LottoPurchasedAmountInput` 테스트를 `ApplicationTest`에서 `InputViewTest`로 분리
- `Buyer` 테스트를 `ApplicationTest`에서 `BuyerTest`로 분리 및 코드 구현에 따른 테스트 코드 변경
