+++ src/test/java/lotto/ApplicationTest.java
@@ -1,10 +1,7 @@
 package lotto;
 
 import camp.nextstep.edu.missionutils.test.NsTest;
-import lotto.dto.LottoPurchasedAmountInput;
 import lotto.exception.ErrorMessage;
-import lotto.exception.LottoException;
-import lotto.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,8 +11,6 @@ import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueN
 import static camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class ApplicationTest extends NsTest {
     private static final String ERROR_MESSAGE = "[ERROR]";
@@ -122,58 +117,6 @@ class ApplicationTest extends NsTest {
             runException("1000j");
             assertThat(output()).contains(ERROR_MESSAGE);
         });
-    }
-
-    @Test
-    void 입력된_문자열이_null_이면_예외가_발생한다() {
-        LottoException exception = assertThrows(LottoException.class, () -> {
-            LottoPurchasedAmountInput.from(null);
-        });
-        assertEquals(ErrorMessage.NULL_INPUT_ERROR.getMessage(), exception.getMessage());
-    }
-
-    @Test
-    void 입력된_문자열이_빈_문자열이면_예외가_발생한다() {
-        LottoException exception = assertThrows(LottoException.class, () -> {
-            LottoPurchasedAmountInput.from("");
-        });
-        assertEquals(ErrorMessage.EMPTY_INPUT_ERROR.getMessage(), exception.getMessage());
-    }
-
-    @Test
-    void 입력된_로또_구입_금액에_숫자가_아닌_문자열이_포홤되면_예외가_발생한다() {
-        assertSimpleTest(() -> {
-            assertThatThrownBy(() -> runException("1200$"))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage(ErrorMessage.PURCHASE_AMOUNT_MUST_BE_NUMBER.getMessage());
-        });
-    }
-
-    @DisplayName("입력된 로또 구입 금액이 1,000원 미만이면 예외가 발생한다.")
-    @Test
-    void 입력된_로또_구입_금액이_1000원_미만이면_예외가_발생한다() {
-        assertSimpleTest(() -> {
-            assertThatThrownBy(() -> runException("900"))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage(ErrorMessage.PURCHASE_AMOUNT_MUST_BE_AT_LEAST_1000.getMessage());
-        });
-    }
-
-    @DisplayName("입력된 로또 구입 금액이 1,000원으로 나누어떨어지지 않으면 예외가 발생한다.")
-    @Test
-    void 입력된_로또_구입_금액이_1000원으로_나누어떨어지지_않으면_예외가_발생한다() {
-        assertSimpleTest(() -> {
-            assertThatThrownBy(() -> runException("1200"))
-                    .isInstanceOf(IllegalArgumentException.class)
-                    .hasMessage(ErrorMessage.PURCHASE_AMOUNT_MUST_BE_MULTIPLE_OF_1000.getMessage());
-        });
-    }
-
-    @Test
-    void 입력된_로또_구입_금액이_int의_범위가_아니면_예외가_발생한다() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, Integer.MAX_VALUE + 1)))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage(ErrorMessage.PURCHASE_AMOUNT_TOO_LARGE.getMessage());
     }
 
     @DisplayName("입력된 보너스 번호가 1~45사이의 숫자가 아니면 예외가 발생한다.")
